feat: Add entitlements endpoints (MPT-4901)
