+++ Makefile
@@ -1,8 +1,13 @@
+ifneq (,$(wildcard ./.env))
+    include .env
+    export
+endif
+
 .env:
 	cp env.example .env
 
 .PHONY: dev-server
-dev-server:
+dev-server: .env
 	uv run fastapi dev app/main.py
 
 .PHONY: lint
@@ -26,3 +31,17 @@ security-checks:
 .PHONY: tests
 tests:
 	uv run pytest
+
+.PHONY: db-cli
+db-cli:
+	pgcli "postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}:${POSTGRES_PORT}/${POSTGRES_DB}"
+
+.PHONY: gen-migration
+gen-migration: message =
+gen-migration:
+	uv run alembic revision --autogenerate -m "$(message)"
+
+
+.PHONY: migrate
+migrate:
+	uv run alembic upgrade head
