+++ env.example
@@ -1,5 +1,5 @@
-POSTGRES_DB: app
-POSTGRES_USER: postgres
-POSTGRES_PASSWORD: mysecurepass
-POSTGRES_HOST: "*"
-POSTGRES_PORT: "5432"
+POSTGRES_DB=postgres
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=mysecurepass
+POSTGRES_HOST=0.0.0.0
+POSTGRES_PORT=5432
