+++ pyproject.toml
@@ -5,17 +5,25 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
+    "alembic>=1.14.0",
+    "asyncpg>=0.30.0",
+    "fastapi-async-sqlalchemy>=0.6.1",
     "fastapi-camelcase>=2.0.0",
     "fastapi[standard]>=0.115.5",
     "pycountry>=24.6.1",
     "pydantic-extra-types>=2.10.0",
+    "pydantic-settings>=2.6.1",
     "pytest-cov>=6.0.0",
+    "python-dotenv>=1.0.1",
+    "sqlalchemy[asyncio]>=2.0.36",
+    "sqlmodel>=0.0.22",
 ]
 
 [dependency-groups]
 dev = [
     "bandit>=1.8.0",
     "httpx>=0.27.2",
+    "ipython>=8.30.0",
     "mypy>=1.13.0",
     "pytest>=8.3.3",
     "pytest-asyncio>=0.24.0",
